+++ Realm/RLMSyncSessionRefreshHandle.mm
@@ -33,7 +33,17 @@ namespace {
 
 void unregisterRefreshHandle(const std::weak_ptr<SyncUser>& user, const std::string& path) {
     if (auto strong_user = user.lock()) {
-        std::static_pointer_cast<CocoaSyncUserContext>(strong_user->binding_context())->unregister_refresh_handle(path);
+        context_for(strong_user).unregister_refresh_handle(path);
+    }
+}
+
+void reportInvalidAccessToken(const std::weak_ptr<SyncUser>& user, NSError *error) {
+    if (auto strong_user = user.lock()) {
+        if (RLMUserErrorReportingBlock block = context_for(strong_user).error_handler()) {
+            RLMSyncUser *theUser = [[RLMSyncUser alloc] initWithSyncUser:std::move(strong_user)];
+            [theUser logOut];
+            block(theUser, error);
+        }
     }
 }
 
@@ -166,11 +176,13 @@ static const NSTimeInterval RLMRefreshBuffer = 10;
 }
 
 /// Handler for network requests that failed before the JSON parsing stage.
-- (BOOL)_handleFailedRequest:(NSError *)error {
+- (void)_handleFailedRequest:(NSError *)error {
     NSError *authError;
     if ([error.domain isEqualToString:RLMSyncAuthErrorDomain]) {
         // Network client may return sync related error
         authError = error;
+        // Try to report this error to the expiration callback.
+        reportInvalidAccessToken(_user, authError);
     } else {
         // Something else went wrong
         authError = make_auth_error_bad_response();
@@ -200,14 +212,14 @@ static const NSTimeInterval RLMRefreshBuffer = 10;
         // This error isn't a network failure error. Just invalidate the refresh handle and stop.
         unregisterRefreshHandle(_user, _path);
         [self invalidate];
-        return NO;
+        return;
     }
     // If we tried to initially bind the session and failed, we'll try again. However, each
     // subsequent attempt will use a weak pointer to avoid prolonging the session's lifetime
     // unnecessarily.
     _strongSession = nullptr;
     [self scheduleRefreshTimer:nextTryDate];
-    return NO;
+    return;
 }
 
 /// Callback handler for network requests.
@@ -222,15 +234,16 @@ static const NSTimeInterval RLMRefreshBuffer = 10;
         // Otherwise, malformed JSON
         unregisterRefreshHandle(_user, _path);
         [self.timer invalidate];
+        NSError *error = make_sync_error(make_auth_error_bad_response(json));
         if (self.completionBlock) {
             self.completionBlock(error);
         }
-        [[RLMSyncManager sharedManager] _fireError:make_sync_error(make_auth_error_bad_response(json))];
-        return NO;
+        [[RLMSyncManager sharedManager] _fireError:error];
     } else {
         REALM_ASSERT(error);
-        return [self _handleFailedRequest:error];
+        [self _handleFailedRequest:error];
     }
+    return NO;
 }
 
 - (void)_timerFired:(__unused NSTimer *)timer {